Updated the localization of the"Region" field label and its dropdowns default value
The updated labels incorporate a space around "/" so that the text wraps on the word. The select value was also made generic for when more than one descriptor applies to the regions.

Lastly the Region label for Canada was updated based on Oxfam's suggestions for including "Territory" explicitly in it.