New Component: UniversalOptIn
In this class, event listeners will be added to all yes/no radio buttons or checkboxes within a universal opt-in element. This element refers to any form block that has the CSS class "universal-opt-in".
When a user clicks on a radio or checkbox button, we will search for all other buttons of the same type within the universal opt-in element and replicate the user's selection.